+++ front-page.php
@@ -10,7 +10,7 @@
 ?>
 <?php get_header(); ?>
 
-<article id="post-<?php the_ID(); ?>" <?php post_class('main-content'); ?>>
+<article id="post-<?php the_ID(); ?>" <?php post_class( 'main-content' ); ?>>
 
     <div class="entry-content">
         <?php