+++ home.php
@@ -12,7 +12,7 @@
 <?php get_header(); ?>
 
 <?php if ( have_posts() ) : ?>
-    <div class="blog-page-posts">
+    <div class="page-posts-grid">
         <?php
         while ( have_posts() ) :
             the_post();