+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # EMdotBike Theme Changelog
 
+## 0.1.4
+
+* Removed custom post header, but support legacy if no block
+* Updated styles to support Gutenberg blocks better
+* Added theme.json file
+
 ## 0.1.31
 
 * Make footer 1 widget and use WP for cols